Add files via upload
Programming Assignment on Reliable-Transport Protocols (Modern Networking Concepts course -university at Buffalo)

CONTENTS OF THIS FILE :
=======================
1. Compilation of Program
2. Validation of protocols
3. Performance comparison
4. The Important Variables Define
	a) Part 1 : defines for abt protocol(abt.c)
	b) Part 2 : defines for abt protocol(gbn.c)
	c) Part 3 : defines for abt protocol(sr.c)
How to compile the program ? ================================ type a simple command "make" as below,
$ make The will create a binary executable of all C code files.

Now run each protocols individually

$ ./abt

$ ./gbn

$ ./sr

Above command will run all protocols.

$ make clean

This will remove all the object file from the directory.

Validation of protocols : ===========================
Run each of your protocols with a total number of 10 messages to be sent by entity A for the Alternating-Bit-Protocol version and 20 messages for the Go-Back-N and the Selective-Repeat versions.Set the loss probability to 0.2, the corruption probability to 0.2, the trace level to 2 and the mean time between messages arrivals (from A’s layer5) to 1000 in the case of the Alternating-Bit-Protocol and 50 in the case of the other two protocols. For the Go-Back-N and the Selective-Repeat versions, set the window size to 10.

Performance comparison : ===========================
In each of the following 2 experiments, run each of your protocols with a total number of 1000 messages to be sent by entity A, a mean time of 50 between message arrivals (from A’s layer5) and a corruption probability of 0.2.

Experiment 1: 

With loss probabilities – {0.1, 0.2, 0.4, 0.6, 0.8}, compare the 3 protocols’ throughputs 
at the application layer of receiver B. Use 2 window sizes - {10, 50} for the 
Go-Back-N version and the Selective-Repeat Version.


Experiment 2: 
With window sizes – {10, 50, 100, 200, 500} for GBN and SR, compare the 3 protocols’ throughputs at the application layer of receiver B. Use 3 loss probabilities – {0.2, 0.5, 0.8} for all 3 protocols.